+++ CHANGELOG.md
@@ -5,7 +5,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-Nothing yet :)
+### Fixed
+- Fixed [issue 35](https://github.com/delightedCrow/WallberryTheme/issues/35) of image not loading due to Unsplash API change.
+- Fixed [issue 33](https://github.com/delightedCrow/WallberryTheme/issues/33) showing incorrect documentation for WallberryTheme config.
+
+### Added
+- Added the following config options to WallberryTheme for customizing images: `imageWidth`, `imageHeight`, and `imageOptions`.
+
+### Removed
+- Removed the `resizeForScreen` config option.
 
 ## 3.0.1 - 2021-1-15
 ### Fixed